# docs/docs.go/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/chats/{chat_id}/messages": {
            "get": {
                "description": "Get messages for a specific chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get messages for chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chat ID",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of messages",
                        "name": "posts_count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timestamp",
                        "name": "ts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forms.MessageOut"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "403": {
                        "description": "User is not a participant in the chat",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            }
        },
...
        "/api/friends": {
            "get": {
                "description": "Возвращает список друзей пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Получить друзей",
                "responses": {
                    "200": {
                        "description": "Список друзей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forms.FriendsInfoOut"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            },
            "post": {
                "description": "Отправляет запрос в друзья конкретному пользователю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Добавить в друзья",
                "responses": {
                    "200": {
                        "description": "Список друзей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/forms.FriendsInfoOut"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "409": {
                        "description": "Отношение между пользователями (подписчик/друг) уже существует",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/forms.ErrorForm"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.ErrorForm": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "forms.FileOut": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "forms.FriendsInfoOut": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_online": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "forms.MessageOut": {
            "type": "object",
            "properties": {
                "audio": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.FileOut"
                    }
                },
                "chat_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.FileOut"
                    }
                },
                "id": {
                    "type": "string"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.FileOut"
                    }
                },
                "sender": {
                    "$ref": "#/definitions/forms.PublicUserInfoOut"
                },
                "stickers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.FileOut"
                    }
                },
                "text": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "forms.PublicUserInfoOut": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "relation": {
                    "$ref": "#/definitions/models.UserRelation"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.UserRelation": {
            "type": "string",
            "enum": [
                "friend",
                "following",
                "followed_by",
                "stranger",
                "self",
                ""
            ],
            "x-enum-varnames": [
                "RelationFriend",
                "RelationFollowing",
                "RelationFollowedBy",
                "RelationStranger",
                "RelationSelf",
                "RelationNone"
            ]
        }
    }
}